NGSTACK-851 delete unused filed, rename project specific stuff, refactor js components
