+++ assets/js/backoffice/index.js
@@ -1,7 +1,5 @@
 import '../../sass/backoffice/style.scss';
-import './sidebar';
 import './components';
-import './ngconversations-mark-thread-as-read-listener';
 import 'bootstrap';
 // eslint-disable-next-line import/no-unresolved
 import 'slim-select/styles';
