--- assets/sass/backoffice/_event_registration.scss
@@ -1,129 +1,0 @@
-.event-session-container {
-    .session-details {
-        padding: 1rem;
-
-        &-row {
-            display: flex;
-            margin-bottom: 16px;
-
-            .details-row-title {
-                flex: 0.4;
-                font-size: 0.875rem;
-                line-height: 1.4;
-                font-weight: 400;
-                color: $secondary_1;
-            }
-
-            .details-row-value {
-                flex: 0.6;
-                font-size: 0.875rem;
-                line-height: 1.4;
-                font-weight: 600;
-                color: $secondary_1;
-
-                a {
-                    height: fit-content;
-                }
-            }
-
-            p {
-                margin: 0;
-            }
-        }
-    }
-
-    .session-files-header {
-        padding: 1rem;
-        border-bottom: 2px solid $secondary_2;
-        border-top: 2px solid $secondary_2;
-        display: grid;
-        grid-template-columns: 1fr auto;
-        grid-template-rows: auto auto;
-
-        h4 {
-            font-size: 1.375rem;
-            line-height: 1.2;
-        }
-
-        p {
-            font-size: 0.75rem;
-            line-height: 1.5;
-            color: rgba($secondary_1, 0.66);
-            margin-bottom: 0;
-        }
-
-        button {
-            grid-row: 1 / 3;
-            grid-column: 2;
-            align-self: center;
-        }
-    }
-
-    .dropdown-menu {
-        i {
-            position: absolute;
-            left: 0;
-            display: flex;
-            padding: 8px;
-            pointer-events: none;
-
-            &::before {
-                content: '';
-                width: 24px;
-                height: 24px;
-                display: inline-block;
-            }
-        }
-
-        i.icon-download::before {
-            @extend %icon-download;
-        }
-
-        i.icon-rename::before {
-            @extend %icon-rename;
-        }
-    }
-
-    ~ .modal {
-        .modal-dialog {
-            .modal-body {
-                .row-input {
-                    display: flex;
-                    flex-direction: column;
-                    margin-bottom: 16px;
-
-                    label {
-                        margin-bottom: 8px;
-                    }
-
-                    input {
-                        width: 100%;
-                    }
-                }
-            }
-        }
-    }
-}
-
-.session-edit-container {
-    form {
-        margin: 1.5rem 1rem;
-        display: flex;
-        flex-direction: column;
-        gap: 1rem;
-
-        .row-input {
-            display: flex;
-
-            label {
-                width: 35%;
-            }
-        }
-
-        .form-actions {
-            margin-left: auto;
-            display: flex;
-            gap: 1rem;
-        }
-    }
-}
