--- assets/sass/backoffice/_event_report.scss
@@ -1,218 +1,0 @@
-.event-report-container {
-    margin: 1rem;
-
-    .filter-form {
-        .report-inputs-container {
-            display: grid;
-            grid-template-columns: 1fr 1fr 1fr;
-            grid-template-rows: 1fr 1fr;
-            gap: 1rem;
-
-            > div {
-                display: flex;
-                flex-direction: column;
-                gap: 0.25rem;
-
-                &:not(.row-input) {
-                    justify-content: flex-end;
-                }
-
-                &:first-child {
-                    grid-column: 1/-1;
-                }
-
-                .ss-main,
-                .flatpickr {
-                    width: 100%;
-                }
-
-                &.compound-date-input {
-                    .form-hidden {
-                        display: none;
-                    }
-
-                    @include media-breakpoint-up(md) {
-                        margin-top: -4px;
-
-                        label {
-                            margin-bottom: 4px;
-                        }
-                    }
-                }
-            }
-        }
-
-        .btn[type='submit'] {
-            // margin-top: 1rem;
-        }
-    }
-}
-
-.event-statistics-container {
-    $item-width: 25rem;
-
-    .event-details-section {
-        padding: 1rem;
-        display: flex;
-        flex-direction: column;
-        gap: 1rem;
-
-        ~ .event-details-section,
-        ~ .charts-form-wrapper .event-details-section {
-            border-top: 2px solid $border;
-        }
-
-        .event-details-section-title {
-            font-size: 1.125rem;
-            font-weight: 600;
-        }
-
-        .event-details-row {
-            display: flex;
-            gap: 1rem;
-
-            .event-details-label {
-                font-size: 0.875rem;
-                font-weight: 400;
-                width: 35%;
-                min-width: 200px;
-                color: $secondary_1;
-            }
-
-            .event-details-value {
-                font-size: 0.875rem;
-                font-weight: 600;
-                color: $secondary_1;
-                width: 65%;
-            }
-
-            .event-details-multiple {
-                width: 65%;
-
-                .event-details-label {
-                    display: inline;
-                    min-width: unset;
-                    width: unset;
-                }
-
-                .event-details-value {
-                    display: inline;
-                    min-width: unset;
-                    width: unset;
-                }
-            }
-        }
-    }
-    .event-statistics-graphical-counts-container {
-        display: flex;
-        gap: 1rem;
-
-        .event-statistics-graphical-count {
-            display: inline-flex;
-            flex-direction: column;
-            gap: 0.5rem;
-            align-items: center;
-            width: $item-width;
-
-            .graphical-count-amount {
-                width: 100%;
-                font-size: 1.5rem;
-                font-weight: 600;
-                display: flex;
-                align-items: center;
-                justify-content: center;
-                height: 15rem;
-                background-color: $secondary_2;
-            }
-
-            .graphical-count-label {
-                font-weight: 600;
-                text-align: center;
-            }
-        }
-
-        + .charts-form-wrapper {
-            margin-top: 2rem;
-        }
-    }
-
-    .charts-form-wrapper {
-        .charts-section-title {
-            font-size: 1.125rem;
-            font-weight: 600;
-        }
-
-        .charts-wrapper {
-            display: flex;
-            gap: 1rem;
-
-            .chart-wrapper {
-                display: inline-flex;
-                flex-direction: column;
-                gap: 0.5rem;
-                align-items: center;
-
-                .chart-count-label {
-                    font-weight: 600;
-                    text-align: center;
-                }
-
-                .chart-count-graph {
-                    background-color: $secondary_2;
-                    padding: 1rem 4rem 2rem;
-                    width: $item-width;
-
-                    canvas {
-                    }
-                }
-            }
-        }
-    }
-
-    .event-statistics-actions-container {
-        margin: 1rem -1rem 0;
-        padding: 1rem 1rem 0 0;
-        display: flex;
-        gap: 1rem;
-        width: calc(100% + 2rem);
-        border-top: 2px solid $border;
-        justify-content: flex-end;
-    }
-}
-
-.create-template-wrapper {
-    padding: 1rem;
-
-    .create-template-title {
-        margin: 0.5rem 0 1rem;
-        font-size: 1.125rem;
-        font-weight: 600;
-        color: $secondary_1;
-    }
-
-    .row-input {
-        display: flex;
-        width: 100%;
-        gap: 1rem;
-
-        label {
-            width: calc(35% - 0.5rem);
-
-            ~ select,
-            ~ input,
-            ~ div {
-                width: calc(65% - 0.5rem);
-            }
-        }
-
-        + .row-input {
-            margin-top: 1rem;
-        }
-    }
-
-    .form-actions-container {
-        display: flex;
-        justify-content: space-between;
-        margin-top: 1rem;
-    }
-}
