+++ assets/sass/backoffice/_header.scss
@@ -85,12 +85,6 @@
                 justify-content: center;
             }
 
-            &.icon-news {
-                &::before {
-                    @extend %icon-news;
-                }
-            }
-
             @media (max-width: 768px) {
                 width: 40px;
                 height: 40px;
