--- assets/sass/backoffice/_technical-enquiry.scss
@@ -1,224 +1,0 @@
-.te-header-info {
-    &-wrap {
-        .datetimepicker {
-            max-width: 400px;
-            border-radius: 4px;
-            height: 40px;
-            margin-bottom: 2rem;
-        }
-    }
-}
-
-.te-survey-wrap {
-    padding: 1rem;
-    .field-group {
-        div.row {
-            .col {
-                padding: 0;
-            }
-        }
-    }
-}
-
-.technical-enquiry-form-wrap {
-    .te-select-expert {
-        margin: 2rem 0;
-
-        .ss-main {
-            height: 40px;
-            margin-right: 1rem;
-        }
-
-        .form-label {
-            margin-bottom: 0;
-        }
-
-        @include media-breakpoint-down(md) {
-            .ngsb-field {
-                width: 100%;
-                margin-bottom: 1rem;
-            }
-        }
-    }
-
-    .te-form-actions {
-        border-top: 2px solid $border;
-        margin: 0 -1rem;
-        padding: 1rem 1rem 0 1rem;
-
-        .te-form-actions-buttons {
-            .mb-3:nth-child(2) {
-                margin-left: auto;
-                margin-right: 0.5rem;
-
-                @include media-breakpoint-down(md) {
-                    margin-left: revert;
-                    margin-right: revert;
-                }
-            }
-
-            @include media-breakpoint-down(md) {
-                flex-direction: column;
-            }
-        }
-    }
-
-    &.ngsb-survey {
-        padding: 1rem 1rem 0 1rem;
-
-        div.cke {
-            width: 100%;
-        }
-
-        .ngsb-field {
-            padding-left: 0;
-            padding-right: 0;
-
-            .row {
-                margin: revert;
-            }
-
-            strong p {
-                font-weight: 600;
-            }
-
-            .field-response {
-                p {
-                    margin-bottom: 0;
-                }
-            }
-
-            .form-label {
-                align-items: center;
-            }
-
-            &.ngsb-field-rich-text {
-                margin-bottom: 1rem;
-                .form-label {
-                    align-items: baseline;
-                }
-            }
-
-            &.ngsb-field-iota-country-organization {
-                div {
-                    .col-form-label {
-                        align-items: center;
-                    }
-                }
-
-                .row .col {
-                    .mb-3 {
-                        .ss-main {
-                            height: 40px;
-                            width: 95%;
-                        }
-                    }
-                }
-            }
-
-            &:not(.ngsb-field-page):not(.ngsb-field-section) {
-                padding: .5rem 0;
-            }
-
-            .field-hint {
-                margin-right: 100%;
-            }
-        }
-
-
-        + .te-view-actions {
-            border-top: 2px solid $border;
-            padding: 1rem;
-
-            @include media-breakpoint-down(md) {
-                .btn {
-                    margin-bottom: 0.5rem;
-                }
-            }
-        }
-
-        @at-root {
-            .ngsb-survey + .te-view-actions {
-                padding: 1rem;
-            }
-        }
-
-    }
-
-    .te-survey-wrap {
-        .ss-main[role="combobox"] {
-            width: 100%;
-            overflow-y: auto;
-        }
-    }
-}
-
-.technical-enquiry.ngsb-survey {
-    .ngsb-field-section {
-        padding: 1rem;
-    }
-
-    .section.row {
-        .ngsb-field-single-choice.col-12[data-field-identifier="answering_countries"],
-        .ngsb-field-iota-member-country.col-12[data-field-identifier="answering_member_countries"] {
-            padding-left: 0;
-        }
-
-        .ngsb-field-single-choice.col-12[data-field-identifier="answering_countries"] {
-            margin-top: 2rem;
-        }
-    }
-
-    + .te-view-actions {
-        padding: 1rem;
-
-        .btn-primary {
-            @include media-breakpoint-down(sm) {
-                margin-bottom: 0.5rem;
-            }
-        }
-    }
-
-    &.ngsb-view-user_response {
-        .ngsb-field:not(.ngsb-field-page):not(.ngsb-field-section) {
-            padding-left: 0.5rem;
-        }
-    }
-}
-
-
-td.answering-countries-column {
-    width: 280px;
-
-    .collapsed {
-        display: none;
-    }
-
-    .collapsed.not-collapsed {
-        display: block;
-    }
-
-    .btn[data-toggle=collapse] {
-        margin: 4px 0;
-    }
-}
-
-td.list-title{
-    a.te-website-link {
-        display: inline-flex;
-        align-items: center;
-        text-decoration: underline;
-        text-decoration-color: $border;
-        text-decoration-thickness: 2px;
-        text-underline-offset: 3px;
-
-        &::after {
-            content: '';
-            @extend %icon-external;
-            display: inline-block;
-            width: 1rem;
-            height: 1rem;
-            margin-left: 4px;
-        }
-    }
-}
