--- assets/sass/backoffice/_user_impersonation.scss
@@ -1,95 +1,0 @@
-.ng-user-impersonation-app {
-    .icon {
-        margin-right: 0.75rem;
-        width: 2rem;
-        height: 2rem;
-
-        &.icon-small{
-            width: 1rem;
-            height: 1rem;
-        }
-    }
-
-    .btn-primary {
-        svg {
-            fill: $white;
-        }
-    }
-}
-
-.ng-user-impersonation-content {
-    .container {
-        max-width: 100%;
-    }
-
-    .users-filter {
-        h3 {
-            font-size: 1.25rem;
-            margin-bottom: 24px;
-        }
-    }
-
-    form {
-        margin: 16px 0;
-
-        > div.row {
-            flex-direction: column;
-            font-size: 0.875rem;
-
-            label {
-                padding-bottom: 8px;
-                padding-top: 0;
-                width: auto;
-            }
-
-            input.form-control,
-            select.form-select {
-                padding-top: 0.4rem;
-                padding-bottom: 0.4rem;
-                border-radius: 4px;
-            }
-        }
-
-        div.col-xs-12 {
-            display: flex;
-
-            a {
-                margin-right: 16px;
-            }
-
-            @include media-breakpoint-down(md) {
-                flex-direction: column;
-
-                a {
-                    margin-bottom: 16px;
-                    margin-right: 0;
-                }
-            }
-        }
-    }
-
-    .users-list {
-        font-size: 0.875rem;
-        margin-top: 32px;
-
-        .section-title {
-            .title {
-                font-size: 20px;
-                margin-bottom: 20px;
-                color: $secondary_4;
-            }
-        }
-
-        .table {
-            tr {
-                td {
-                    vertical-align: middle;
-                }
-
-                ul {
-                    margin-bottom: 0;
-                }
-            }
-        }
-    }
-}
