+++ assets/sass/backoffice/style.scss
@@ -29,23 +29,11 @@
 @import 'main';
 @import 'filter_form';
 @import 'listing';
-@import 'event_hosting';
-@import 'event_report';
 @import 'top_navigation';
-@import 'survey';
-@import 'user_impersonation';
-@import 'event_registration';
 @import 'spinner';
 @import 'modal';
-@import 'my_country';
 @import 'my_account';
-@import 'news';
-@import 'create';
-@import 'event';
-@import 'mailing';
-@import 'technical-enquiry';
 @import 'user_profile';
-@import 'cbip';
 @import 'global_styles';
 
 @import 'print';
