+++ templates/themes/app/backoffice/layout.html.twig
@@ -54,12 +54,6 @@
             <span class="hamburger"></span>
         </a>
 
-{#        <a href="{{ ibexa_path(ibexa.rootLocation) }}" class="app-header-logo">#}
-{#            {% if not ngsite.siteInfoContent.fields.site_logo_inverted.empty %}#}
-{#                <img src="{{ asset(ngsite.siteInfoContent.fields.site_logo_inverted.value.id, 'ibexa_file') }}" alt="{{ ngsite.siteInfoContent.fields.site_name.value.text|trim }}" />#}
-{#            {% endif %}#}
-{#        </a>#}
-
         <div class="app-header__content">
             {% block header %}
                 <div class="app-header__content-wrap">
