+++ templates/themes/app/backoffice/main_menu.html.twig
@@ -53,12 +53,6 @@
                 <ul>
                     <li><a href="">{{ 'menu.main_menu.personal_details'|trans }}</a></li>
                     <li><a href="">{{ 'menu.main_menu.user_credentials'|trans }}</a></li>
-
-                    {% if is_granted('ibexa:netgen_ibexa_user_impersonation:impersonate') %}
-                        <li><a href="{{ path('netgen.ibexa_user_impersonation.admin.list') }}">{{ 'menu.main_menu.impersonate'|trans }}</a></li>
-                    {% elseif is_granted('ROLE_IMPERSONATED_USER') %}
-                        <li><a href="{{ path('netgen.ibexa_user_impersonation.app.user.switch_back') }}">{{ 'layout.user_settings.end_impersonation'|trans }}</a></li>
-                    {% endif %}
                 </ul>
             </div>
             <div class="user-settings-menu-footer">
